% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_to_gps.R
\name{address_to_gps}
\alias{address_to_gps}
\title{Google api call for turning address to gps coordinates}
\usage{
address_to_gps(address)
}
\arguments{
\item{address}{string with address to convert}
}
\value{
data.frame with GPS coordinates
}
\description{
Uses google's api from \code{ggmap} to convert address to gps
}
\examples{
 address <- "31 Pickwick street Salt River Western Cape South Africa"
 address_to_gps(address)
}
