% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_plot.R
\name{calendar_plot}
\alias{calendar_plot}
\title{Time Series Calendar Heatmap plot}
\usage{
calendar_plot(data, date, n, low = "#a7b5c3", high = "red")
}
\arguments{
\item{data}{data containing the values you want to plot}

\item{date}{name of date column}

\item{n}{name of column with the numerical amount to grade plot}

\item{low}{color on low gradient}

\item{high}{color of high gradient}
}
\value{
calendar heatmap plot
}
\description{
Utilizes ggplot's tile geom to construct heatmaps for calendar dates
}
\examples{

library(quantmod)
library(dplyr)
library(tbl2xts)

getSymbols("JSE:WHL", src = "google")

stock <- `JSE:WHL`

stock <- xts_tbl(stock)
  
stock \%>\% filter(date > "2016-01-01") \%>\% 
  calendar_plot(date, JSE.WHL.Volume)

stock \%>\% filter(date > "2016-01-01") \%>\% 
  calendar_plot(date, JSE.WHL.Volume, low = "#a7b5c3")


}
