% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_plot.R
\name{dual_plot}
\alias{dual_plot}
\title{Plotting of 2 series with a secondary Y axis}
\usage{
dual_plot(x1 = time(y1), y1, y2, x2 = x1, col1 = "#C54E6D",
  col2 = "#009380", lwd = 1, colgrid = NULL, mar = c(3, 6, 3, 6) + 0.1,
  ylab1 = substitute(y1), ylab2 = substitute(y2), nxbreaks = 5,
  yleg1 = paste(ylab1, "(LHS)"), yleg2 = paste(ylab2, "(RHS)"),
  ylim1 = NULL, ylim2 = NULL, ylim.ref = 1, xlab = "", main = NULL,
  legx = "topleft", legy = NULL, ...)
}
\arguments{
\item{x1}{x-axis object}

\item{y1, y2}{objects to be plotted}

\item{col1, col2}{colors for the 2 lines}

\item{lwd}{line width}

\item{colgrid}{colour of gridlines; if NULL there are no gridlines}

\item{mar}{margins for plotting area see  (see ?par)}

\item{ylab1, ylab2}{labels for the two y axes}

\item{yleg1, yleg2}{legend of the 2 objects}

\item{legx, legy}{position of the legend}

\item{nbreaks}{Number of breaks in horizontal axis}
}
\value{
Plot of 2 series on different Y-axis
}
\description{
dual_plot will take 2 series and plot the second series on the second Y-axis
}
\examples{
data("forex")

with(forex, dual_plot(x1 = TimePeriod, y1 = NZDUSD, y2 = TWI, lwd = 1.2, colgrid = "grey90", 
    main = "NZ dollar exchange rate & trade-weighted index",
    ylab1 = "NZ dollars for one US dollar",
    ylab2 = "Index",
    yleg1 = "NZD / USD exchange rate (left axis)",
    yleg2 = "Trade-weighted index (right axis)",
    mar = c(5,6,3,6)))
    
mtext(side = 1, "Data from RBNZ", 
          adj = 1, cex = 0.8, line = 3)
          
}
