% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_forex.R
\docType{data}
\name{forex}
\alias{forex}
\title{Time series of the Forex and TWI}
\format{a \code{data frame} with 391 rows and 3 variables
\describe{
\item{NZDUSD}{exchange rate, NZD to USD}
\item{TWI}{Trade-weighted Index}
\item{TimePeriod}{Time period}
}}
\source{
\url{http://www.rbnz.govt.nz/statistics/key-graphs/key-graph-exchange-rate}
}
\usage{
data(forex)
}
\description{
A dataset containg the exchange rate for the 
New Zealand Dollar against USD. As well as the 
Trade-weighted Index for New Zealand
}
\references{
\url{http://ellisp.github.io/blog/2016/08/18/dualaxes}
}
\keyword{datasets}
