% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkgs.R
\name{load_pkg}
\alias{load_pkg}
\title{Easy load operator for packages}
\usage{
load_pkg(packagelist)
}
\arguments{
\item{packagelist}{Vector of packages to load into R}
}
\value{
Packages loaded into R
}
\description{
load_pkg Loads a list of packages. If a package requires installation, the function will install it from CRAN.
}
\examples{
load_pkg(c("ggplot2", "dplyr"))
}
