% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_toolbox.R
\name{load_toolbox}
\alias{load_toolbox}
\title{Easy loading all analytics packages}
\usage{
load_toolbox(install = T)
}
\arguments{
\item{install}{Will install all packages if not available to load. Default = TRUE}
}
\value{
Packages loaded into R for analytics
}
\description{
load_toolbox Loads a list of packages that are most often used in our analytics workspace: 
dplyr, readr, lubridate, tidyr, openxlsx, ggplot, purrr
}
\examples{
load_toolbox()
}

