% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa_data_reweight.R
\name{mfa_data_reweight}
\alias{mfa_data_reweight}
\title{Reweigh dataset based on MFA results}
\usage{
mfa_data_reweight(dataset, mfa_res, index, dim, plot = T)
}
\arguments{
\item{dataset}{the dataset that needs to be reweighted}

\item{mfa_res}{a MFA object that contains the dimension reduction results}

\item{index}{if index is provided it will be binded to dataframe in the end}

\item{dim}{number of dimensions to keep}

\item{plot_scree}{plots the eigenvalues against the factors for interpretation}
}
\value{
MFA Object
}
\description{
uses the results from the multiple factor analysis to reweigh all data based on sample results
}
\examples{

load_toolbox()

data(sales)

dataset <- sales
dataset[,1] <- factor(dataset[,1])

# Samle data and run MFA on the sample
sample <- sample_n(v, 1000)
 
load_pkg("FactoMineR")
mfa_res <- parMFA(sample,
                   group=c(1, 4, 12), 
                   type = c('n','s','s'), 
                   ncp=12, 
                   name.group=c('factor', 'purchases','seasonality'),
                   graph=FALSE, cores = 3)
                   
# The dimension needs to be specified
mfa_data_reweight(dataset, mfa_res)

mfa_data_reweight(dataset, mfa_res, dim = 12)
 
}

