% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa_eval_plots.R
\name{mfa_eval_plots}
\alias{mfa_eval_plots}
\title{Wrapper function for factorextra plotting functions}
\usage{
mfa_eval_plots(mfa_res, indi = F)
}
\arguments{
\item{mfa_res}{an MFA object}

\item{indi}{also plotting the individual results - avoid if data is large!}
}
\value{
Plots of MFA evaluations
}
\description{
Leverages factoextras plotting functions to plot MFA results. Mostly the global.pca object
}
\examples{
library(FactoMineR)

 data(wine)
 res <- MFA(wine, group=c(2,5,3,10,9,2), type=c("n",rep("s",5)),
            ncp=5, name.group=c("orig","olf","vis","olfag","gust","ens"),
            num.group.sup=c(1,6))
 
 mfa_eval_plots(res, indi = T)
}
