% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_action.R
\name{na_action}
\alias{na_action}
\title{Counting/NAs of NA's per column or row vector}
\usage{
na_action(df, margin = 2, replace, drop = F, cutoff = 1)
}
\arguments{
\item{df}{object to be checked}

\item{margin}{Margin sets whether you count are counting columns, margin = 2, or rows ,margin =1 (Default = 2)}

\item{drop}{If this is TRUE, then all rows with NAs will be removed}

\item{cutoff}{Cutoff is a parameter where the exact NA cutoff per row must be when filtering the dataset (Default = 1)}
}
\value{
An count of NAs
}
\description{
na_action Looks at your data.frame() or vector and counts the NAs. Option to remove is also built in
}
\examples{
x <- data.frame(NR = c(1,2,3,5,NA,6,7,NA),Letters = c(rep(LETTERS)[1:7],NA), stringsAsFactors = F)
# Get the count per columns
na_action(x, 2)

# Get the count per row
na_action(x, 1)

# Returns NA count in the vector
na_action(c(1,2,3,NA,4,NA,5))

# Replace all NA with 0

na_action(df, replace = 0)

# Drop all rows with NA
na_action(x, drop =T)

# Specify cutoff for dropoff. 
# Specifying cutoff as 2 will only drop rows where there are more than >=2 NAs in the row
na_action(x, drop = T, cutoff = 2)

# Cutoffs can also be proportional
na_action(x, drop = T, cutoff = 0.6)
 
}
