% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative_match.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Operator that produces negative match vector}
\usage{
x \%!in\% y
}
\arguments{
\item{x, y}{variables to match on}
}
\value{
negative match vector
}
\description{
Negative match function
}

