% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytnyt.R
\name{nytnyt}
\alias{nytnyt}
\title{wrapper for \code{Sys.sleep}}
\usage{
nytnyt(periods = c(2, 3))
}
\arguments{
\item{periods}{controls the min and max for \code{runif} that controls the time for sleep}
}
\value{
\code{Sys.sleep}
}
\description{
A wrapper that is useful when you need to sleep in a loop or scrape
}
\examples{
nytnyt(periods = c(1,2))
}
