% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_xlsx.R
\name{open_xlsx}
\alias{open_xlsx}
\title{Using openxlsx to Create or Update an Excel Workbook}
\usage{
open_xlsx(nameVector, tableList, workBookName, colNames = TRUE,
  rowNames = FALSE, loadWorkBook = FALSE, replaceSheets = FALSE)
}
\arguments{
\item{nameVector}{a single string or a vector of strings}

\item{tableList}{a single data frame or a list of data frames}

\item{workBookName}{a string}

\item{colNames}{a boolean; default = TRUE}

\item{rowNames}{a boolean; default = FALSE}

\item{loadWorkBook}{a boolean; default = FALSE}

\item{replaceSheets}{a boolean; default = FALSE}
}
\value{
.xlsx file
}
\description{
open_xlsx will take a single data frame or a list of data frames, 
   as well as a vector of worksheet names and a workbook name, 
   to create or update  an excel xlsx file with each data frame 
   on a seperate worksheet within the file
}
\examples{
 
# Name the sheets
nameVec1 <- c("1", "2", "3")
nameVec2 <- c("4", "5", "6")
nameVec3 <- c("3", "5", "7")

a <- data.frame(1:3, 4:6, 7:9)
b <- data.frame(10:12, 13:15, 16:18)
tableList1 <- list(a, a, a)
tableList2 <- list(b, b, a)

open_xlsx("1", a, "test")
open_xlsx(nameVec1, tableList1, "test")
open_xlsx(nameVec2, tableList1, "test", loadWorkBook = TRUE)
open_xlsx(nameVec3, tableList2, "test", loadWorkBook = TRUE, replaceSheets = TRUE)

file.remove("test.xlsx")
}

