% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parKmeans.R
\name{parKmeans}
\alias{parKmeans}
\title{Parallel implementation of Kmeans}
\usage{
parKmeans(dataset, seeds, centers, cores = 10,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen")[1])
}
\arguments{
\item{dataset}{the dataset upon which the kmeans algoritm will be run}

\item{seeds}{the number of seed you would like to test for a global min}

\item{centers}{vector of k centers to try}

\item{cores}{number of cores to run in parallel, default is 10}
}
\value{
Data frame with each of the runs' metrics and seed number
}
\description{
Parallel implementation of Kmeans to test for global minimum
}
\examples{

data("sales")

# remove NAs
seeds <- 50
centers <- c(5:15)

best_fit <- parKmeans(sales, seeds = seeds,  centers = centers)
best_fit

# plot runs
load_pkg("ggplot2")

best_fit \%>\% select(clusters, withinss, seed) \%>\% mutate(seed = factor(seed)) \%>\%
  ggplot(aes(x = clusters, y = withinss, color = seed)) +
  geom_line(show.legend=FALSE) +
  geom_smooth(se = F, color = "black", size = 1, span = 0.6)  + 
  ylab(label="WithinSS") + 
  xlab("Clusters") + 
  theme_bw() +
  scale_colour_manual(values=c(rep("#8599bc", seeds)))

}
