% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rf_tree.R
\name{plot_rf_tree}
\alias{plot_rf_tree}
\title{Plot Random Forest trees}
\usage{
plot_rf_tree(model_rf, spec = c("least", "most")[1])
}
\arguments{
\item{final_model}{object of class randomForest from \link[randomForest]{caret}}

\item{tree_num}{tree number you want to plot}
}
\value{
ggplot object of a random forest containing least or most nodes
}
\description{
Plots least and most node trees from Random Forest object
}
\examples{

library(caret)


set.seed(42)
index <- createDataPartition(iris$Species, p = 0.7, list = FALSE)
train_data <- iris[index, ]
test_data  <- iris[-index, ]

fitControl <- trainControl(method = "repeatedcv", 
             number = 5, 
             repeats = 2,
             sampling = "smote",
             savePredictions = TRUE, 
             verboseIter = FALSE)

rfGrid <-  expand.grid(mtry = c(2: sqrt(ncol(train_data))))
rfGrid <-  expand.grid(mtry = c(2: 4))

# run model
set.seed(42)
model_rf <- train(Species ~ .,
                         data = train_data,
                         method = "rf",
                         preProcess = c("scale", "center"),
                         fitControl = fitControl,
                         tuneGrid = rfGrid)

plot(model_rf)
ggplot(model_rf)
plot(model_rf$finalModel)

plot_rf_tree(model_rf, "most")
plot_rf_tree(model_rf, "least")



}
