% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_count.R
\name{prop_count}
\alias{prop_count}
\title{Gives a proportional count of variables supplied}
\usage{
prop_count(df, ...)
}
\arguments{
\item{df}{a \code{tbl} of data}

\item{...}{variables to group by}
}
\description{
Gives a proportional count of variables supplied
}
\examples{
dplyr::starwars \%>\% prop_count(homeworld, species)
}
