% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recessions.R
\docType{data}
\name{recessions}
\alias{recessions}
\title{Time series of the recession data from 1857}
\format{a \code{data frame} with 33 rows and 2 variables
\describe{
\item{Peak}{Peak of recession}
\item{Trough}{Peak of recession}
}}
\usage{
data(recessions)
}
\description{
A dataset containg the exchange rate for the 
New Zealand Dollar against USD. As well as the 
Trade-weighted Index for New Zealand
}
\references{
\url{ http://research.stlouisfed.org/fred2/help-faq/}
}
\keyword{datasets}
