% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Cross section sales analysis per individual}
\format{a \code{data frame} with 20000 rows and 17 variables}
\usage{
data(sales)
}
\description{
A dataset contain various sales metrics explaining customer behaviour
}
\keyword{datasets}
