% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_Publication.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Theme to be used in journal publications}
\usage{
scale_colour_Publication(...)
}
\value{
NA
}
\description{
Theme for use in journals
}
