% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_Publication.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Theme to be used in journal publications}
\usage{
scale_fill_Publication(...)
}
\value{
NA
}
\description{
Theme for use in journals
}
