% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_method.R
\name{show_method}
\alias{show_method}
\title{Get method}
\usage{
show_method(obj, sleep = 1)
}
\arguments{
\item{obj}{names of obj}

\item{sleep}{sleep for seconds before display result}
}
\value{
vector of column names
}
\description{
Get methods of S3 class objects from anywhere
}
\examples{
xx <-  data.frame()
get_method(xx)
}
