% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitplot.R
\name{splitplot}
\alias{splitplot}
\title{Plot dimension description visually}
\usage{
splitplot(mfa, dim = 6, type = c("combined", "quali", "quanti")[3],
  grid = c(2, 3), specific = F)
}
\arguments{
\item{mfa}{MFA object from FactoMineR}

\item{dim}{number of dimensions to plot}

\item{type}{type of plot to return}

\item{grid}{grid to use for plotting (default is c(2,3))}

\item{specific}{whether you are plotting specific dimensions and not the range of dimensions}
}
\value{
Plot containing correlations and v.tests
}
\description{
Function takes MFA object and plots either the quali, quanti or combined metrics of the dimension description
}
\examples{
 
load_pkg("FactoMineR")

data(wine)

mfa <- MFA(wine, group=c(2,5,3,10,9,2), type=c("n",rep("s",5)),
          ncp=5, name.group=c("orig","olf","vis","olfag","gust","ens"),
          num.group.sup=c(1,6), graph = F)

splitplot(mfa, dim = 5, type = "quanti")

# Wine example does not contain any qualitative examples
#splitplot(mfa, dim = 5, type = "quali")

splitplot(mfa, dim = 5, type = "combined")

splitplot(mfa, dim = c(1:2), type = "combined", grid = c(1,2), specific = T)

}
