% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Theme to be used in journal publications}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{size of text in the graph}

\item{base_family}{font of text in graph}
}
\value{
NA
}
\description{
Theme for use in journals
}
\examples{

library(ggplot2)
library(gridExtra)
library(reshape2)

Scatter <-
 ggplot(mtcars, aes(mpg, disp, color = factor(carb))) + 
 geom_point(size = 3) + 
 labs(title = "Scatter Plot")

grid.arrange(Scatter,
            (Scatter + scale_colour_Publication() + theme_Publication()), 
            nrow = 1)

Bar <- ggplot(mtcars, aes(factor(carb), fill = factor(carb))) +
 geom_bar(alpha = 0.7) +
 labs(title = "Bar Plot")

grid.arrange(Bar,
            (Bar + scale_fill_Publication() + theme_Publication()),
            nrow = 1)

Bubble <- ggplot(mtcars, aes(mpg, disp, color = factor(carb), size = hp)) +
 geom_point(alpha = 0.7) +
 labs(title = "Bubble Plot") +
 scale_size_continuous(range = c(3, 10))

grid.arrange(Bubble,
            (Bubble + scale_colour_Publication() + theme_Publication()),
            nrow = 1)


mtcars$Index <- 1:nrow(mtcars)
dat <-
 melt(mtcars,
      id.vars = c("Index"),
      measure.vars = c("drat", "wt"))
Line <- ggplot(dat, aes(Index, value, colour = variable)) +
 geom_line(size = 1.3) +
 labs(title = "Line Plot")
grid.arrange(Line,
            (Line + scale_colour_Publication() + theme_Publication()),
            nrow = 1)


}
