% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_catdesc.R
\name{tidy_catdesc}
\alias{tidy_catdesc}
\title{Tidy Catdesc}
\usage{
tidy_catdesc(df, cluster)
}
\arguments{
\item{df}{data frame to analyse}

\item{cluster}{cluster column name}
}
\value{
list object
}
\description{
Uses FactoMineR::catdesc function to describe the categories of one factor by categorical variables and/or by quantitative variables
}
\examples{

library(FactoMineR)
data(iris)
# Principal Component Analysis:
res.pca <- PCA(iris[,1:4], graph=FALSE)
# Clustering, auto nb of clusters:
hc <- HCPC(res.pca, nb.clust=-1)

hc$data.clust \%>\% tidy_catdesc(., clust)
hc \%>\% tidy_catdesc()
}
