% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_hcpc.R
\name{tidy_hcpc}
\alias{tidy_hcpc}
\title{Tidy wrapper for HCPC results}
\usage{
tidy_hcpc(hcpc_res)
}
\arguments{
\item{hcpc_res}{object of class HCPC from \link[FactoMineR]{HCPC}}
}
\value{
list object containing most important quali and quanti variables in a clust
}
\description{
Uses HCPC class as inputs and tidy's the quali and quanti results of the most important variable in a cluster
}
\examples{

 library(FactoMineR)
 
 data(wine)
 mfa_res <- MFA(wine, group=c(2,5,3,10,9,2), type=c("n",rep("s",5)),
                ncp=5, name.group=c("orig","olf","vis","olfag","gust","ens"),
                num.group.sup=c(1,6))
 
 hcpc_res <- HCPC(res, nb.clust = 3)
 
 tidy_hcpc(hcpc_res)


}
