% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mfa_dim.R
\name{tidy_mfa_dim}
\alias{tidy_mfa_dim}
\title{Tidy MFA global PCA dim ("coord", "cor", "cos2", "contrib") results}
\usage{
tidy_mfa_dim(mfa_res, n = 5, type = c("coord", "cor", "cos2", "contrib")[4],
  top_bottom_n = 5)
}
\arguments{
\item{mfa_res}{object of class MFA}

\item{n}{number of dimension to tidy}

\item{type}{metric to tidy}

\item{top_bottom_n}{decide on top n variables in dimension}
}
\value{
Tidy data frame for MFA global PCA results
}
\description{
Takes an MFA objects and tidy the data into a much easier working format for evaluation
}
\examples{
 library(FactoMineR)

 data(wine)
 res <- MFA(wine, group=c(2,5,3,10,9,2), type=c("n",rep("s",5)),
            ncp=5, name.group=c("orig","olf","vis","olfag","gust","ens"),
            num.group.sup=c(1,6))
 
 tidy_mfa_dim(res)
 
 data (poison)
 MFA(poison, group=c(2,2,5,6), type=c("s","n","n","n"),
     name.group=c("desc","desc2","symptom","eat"),
     num.group.sup=1:2)
 
 tidy_mfa_dim(res)

}

